refactor reaction filtering, move from CardCollection to Room header
